get_average_vecs -> iterator's func
