+++ main.c
@@ -5,12 +5,11 @@
 int	main(void)
 {
 	t_draw		draw;
-	t_vector	ave_vecs;
 
 	draw.map_size = W_MAP * H_MAP;
 	draw.vecs = string_to_vectors(read_file("./torus.3d"), &(draw.vecs_size));
-	ave_vecs = get_average_vectors(draw.vecs, draw.vecs_size);
-	matrix_translate(draw.matrix4, -1 * ave_vecs.x, 0, -1 * ave_vecs.z);
+	vecs_iter(&draw, get_average_vecs);
+	matrix_translate(draw.matrix4, -1 * draw.vecs_ave.x, 0, -1 * draw.vecs_ave.z);
 	vecs_iter(&draw, do_matrix);
 	while (1)
 	{
