+++ utils.c
@@ -1,6 +1,5 @@
 #include "term3d.h"
 #include <math.h>
-#include <string.h>
 
 #ifndef M_PI
 # define M_PI 3.141592653589793
@@ -9,20 +8,4 @@
 double	to_radians(double degrees)
 {
 	return (degrees * M_PI / 180);
-}
-
-void	do_matrix(t_vector *v, t_draw *d)
-{
-	*v = affine(d->matrix4, v);
-}
-
-void	mapping(t_vector *v, t_draw *d)
-{
-	t_vector	tmp;
-	int			pixel;
-
-	tmp = affine(d->matrix4, v);
-	pixel = convert_to_mapindex(&tmp);
-	if (0 <= pixel && pixel <= d->map_size && d->map[pixel] < strlen(ASCII) - 1)
-		d->map[pixel]++;
 }
